% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coarsegrainMultivariate.R
\name{coarsegrainMultivariate}
\alias{coarsegrainMultivariate}
\title{Coarsegrain a multivariate signal}
\usage{
coarsegrainMultivariate(mat, eps, fun = mean, ...)
}
\arguments{
\item{eps}{The bin size.}

\item{fun}{The function to apply to each bun (default = mean).}

\item{...}{Other arguments passed to fun.}

\item{x}{A numeric matrix with dimensions source x time.}
}
\description{
Function coarsegrains a multivariate signal by partitioning the signal into bins
of length eps, and then applying a function (by default, the mean) to each bin.
}
