% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coarsegrainUnivariate.R
\name{coarsegrainUnivariate}
\alias{coarsegrainUnivariate}
\title{Coarsegrain a univariate signal}
\usage{
coarsegrainUnivariate(x, eps, fun = mean, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{eps}{The bin size.}

\item{fun}{The function to apply to each bun (default = mean).}

\item{...}{Other arguments passed to fun.}
}
\description{
Function coarsegrains a univariate signal by partitioning the signal into bins
of length eps, and then applying a function (by default, the mean) to each bin.
}
