% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSMVDiffEn.R
\name{MSMVDiffEn}
\alias{MSMVDiffEn}
\title{Multiscale multivariate sample entropy}
\usage{
MSMVDiffEn(
  mat,
  M,
  tau,
  eps = 1,
  breaks = 100,
  scaleMat = TRUE,
  fun = mean,
  ...
)
}
\arguments{
\item{mat}{A pxn numeric matrix containing a p-variate time series.}

\item{M}{A vector of length p specifying the embedding dimension.}

\item{tau}{A vector of length p specifying the time lag parameter.}

\item{eps}{An integer specifying the time scale.}

\item{fun}{Function applied to each bin, passed to coarsegrain (default = mean).}

\item{...}{Other arguments passed to fun.}

\item{r}{The similarity threshold.}

\item{scale}{If TRUE (default), standardizes the rows of mat.}
}
\description{
Wrapper for MVSampEn that allows the user to specify an integer scale
parameter eps. The time series is coarsegrained, and then the entropy is
computed for the specified time scale. Currently being tested, and is not exported.
}
