% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVDiffEn.R
\name{MVDiffEn}
\alias{MVDiffEn}
\title{Multivariate sample entropy}
\usage{
MVDiffEn(mat, M, tau, breaks = 100, scaleMat = TRUE)
}
\arguments{
\item{mat}{A pxn numeric matrix containing a p-variate time series.}

\item{M}{A vector of length p specifying the embedding dimension.}

\item{tau}{A vector of length p specifying the time lag parameter.}

\item{r}{The similarity threshold.}

\item{scale}{If TRUE (default), standardizes the rows of mat.}

\item{fun}{Function applied to each bin, passed to coarsegrain (default = mean).}

\item{...}{Other arguments passed to fun.}
}
\description{
Wrapper for MVSampEn that allows the user to specify an integer scale
parameter eps. The time series is coarsegrained, and then the entropy is
computed for the specified time scale. Currently being tested, and is not exported.
}
