% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSampEn.R
\name{MVSampEn}
\alias{MVSampEn}
\title{Multivariate sample entropy}
\usage{
MVSampEn(mat, M, tau, r, scale = TRUE)
}
\arguments{
\item{mat}{A pxn numeric matrix containing a p-variate time series.}

\item{M}{A vector of length p specifying the embedding dimension.}

\item{tau}{A vector of length p specifying the time lag parameter.}

\item{r}{The similarity threshold.}

\item{scale}{If TRUE (default), standardizes the rows of mat.}
}
\description{
Function computes the multivariate sample entropy of a signal
}
